Move LoadBalancerServiceInstanceCookieTransformer to loadbalancer module.
